+++ src/driver/integer_task.ml
@@ -678,6 +678,8 @@ let execution_commands task span ~active =
           publication.function_header.declaration_command
       | Function_position_written p ->
           p.position_function.function_header.declaration_command
+      | Function_local_allocated p ->
+          p.allocation_function.function_header.declaration_command
       | Function_parameter_declared p ->
           p.parameter_function.function_header.declaration_command
       | Function_parameter_completed p ->