+++ src/sema/source_activation.ml
@@ -153,6 +153,8 @@ let event_context = function
         | Parser.Function_declared p -> p.function_header.declaration_command
         | Parser.Function_position_written p ->
             p.position_function.function_header.declaration_command
+        | Parser.Function_local_allocated p ->
+            p.allocation_function.function_header.declaration_command
         | Parser.Function_parameter_declared p ->
             p.parameter_function.function_header.declaration_command
         | Parser.Function_parameter_completed p ->